lib: rename files for old mixer implementation
