+++ src/clover_sketches_lib/lib/mixer_legacy/draw_mixer_000.hpp
@@ -4,7 +4,7 @@
 // Copyright (C) 2025  Rob W. Albus
 // Licensed under the GPLv3. See LICENSE for details.
 
-#include "mixer.hpp"
+#include "mixer_000.hpp"
 
 // void mixer_component(const char* id, composition_mix* mixer);
 void draw_mixer_000(const char* id, mixer_list_000* mixer);
