+++ src/clover_sketches_lib/lib/mixer_legacy/mixer_000.cpp
@@ -5,7 +5,7 @@
 #include <format>
 #include <string>
 
-#include "mixer.hpp"
+#include "mixer_000.hpp"
 
 mixer_map_000 build_mixer_map(mixer_list_000& mixer_tracks) {
     mixer_map_000 audio_mixer;
