+++ src/sketches/022_sd_reso_chords/composition/patches/patches.hpp
@@ -7,7 +7,7 @@
 #include "lib/echo/echo.hpp"
 #include "lib/fdn/fdn8_000.hpp"
 #include "lib/kick_drum/kick_drum.hpp"
-#include "lib/mixer/mixer.hpp"
+#include "lib/mixer_legacy/mixer_000.hpp"
 #include "lib/peq/peq.hpp"
 #include "lib/subtractive_synth/subtractive_synth.hpp"
