+++ src/sketches/023_fdn/gui/controller/controllers.cpp
@@ -10,7 +10,7 @@
 #include "lib/fdn/draw_fdn8_023.hpp"
 #include "lib/kick_drum/draw_kick_drum.hpp"
 #include "lib/logging/draw_gui_log_canvas.hpp"
-#include "lib/mixer/draw_mixer.hpp"
+#include "lib/mixer_legacy/draw_mixer_000.hpp"
 #include "lib/peq/draw_peq.hpp"
 
 #include "context.hpp"
