+++ src/sketches/024_dub/graph/graph.hpp
@@ -11,7 +11,7 @@
 #include "lib/cymbal/cymbal_024.hpp"
 #include "lib/fdn/fdn8_023.hpp"
 #include "lib/kick_drum/kick_drum.hpp"
-#include "lib/mixer/mixer.hpp"
+#include "lib/mixer_legacy/mixer_000.hpp"
 #include "lib/peq/peq.hpp"
 #include "lib/subtractive_synth/subtractive_synth.hpp"
