+++ src/sketches/029_app/graph/patches/patches.hpp
@@ -7,7 +7,7 @@
 #include "lib/cymbal/cymbal_000.hpp"
 #include "lib/fdn/fdn8_023.hpp"
 #include "lib/kick_drum/kick_drum.hpp"
-#include "lib/mixer/mixer.hpp"
+#include "lib/mixer_legacy/mixer_000.hpp"
 #include "lib/peq/peq.hpp"
 #include "lib/subtractive_synth/subtractive_synth.hpp"
