+++ src/sketches/031_reso_snare/graph/graph.cpp
@@ -8,7 +8,7 @@
 #include "clover/dsp/wave.hpp"
 #include "clover/math.hpp"
 
-#include "lib/mixer/mixer.hpp"
+#include "lib/mixer_legacy/mixer_000.hpp"
 
 #include "graph.hpp"
 #include "patches/patches.hpp"
