+++ src/sketches/032_meter/gui/controller/controllers.cpp
@@ -9,7 +9,7 @@
 #include "lib/fdn/draw_fdn8_023.hpp"
 #include "lib/kick_drum/draw_kick_drum.hpp"
 #include "lib/meter/draw_meter.hpp"
-#include "lib/mixer/draw_mixer.hpp"
+#include "lib/mixer_legacy/draw_mixer_000.hpp"
 #include "lib/peq/gpeq.hpp"
 
 #include "graph/graph.hpp"
